+++ CHANGES.rst
@@ -3,7 +3,7 @@
 Changes
 #######
 
-0.1.71 (unreleased)
+0.1.71 (2024-09-12)
 ===================
 
 Backwards incompatible: